Add option to write outline information to a file with --dump-outline FILE
The file will be structured as follows:

The first line contains the text "Pages: ", followed by the total
number of pages in the PDF (all numbers are in ascii decimal digits).

Each following line looks like this

 "Level: " <level> "  Page: "  <page number> "  Title: " <title> "\n"

where <level> is an integer indicating the significance of the
heading, <page number> is self-explanatory, and <title> is the
title encoded as utf-8 text that has been escaped as in a URI:
non-alphanumeric characters are replaced by "%" followed by two
hexadecimal digits giving their value ("percent-encoding" as used
in urls and described in section 2.1 of RFC 3986).